[BUGFIX] Apply initial value to additional form element property paths
Adding a new validator to a form element in form editor led to an
error if the validator defines additionalElementPropertyPaths. This
happened because the initial value of those property paths was not
applied to the form element when the validator was added. Therefore, it
was not possible to unset those values as they did not exist unless
their initial value has changed.

With this patch applied, the initial value of validator options is now
passed to all properties of additionalElementPropertyPaths as well.

Resolves: #97040
Releases: main, 11.5
Change-Id: Ife242a48276789fcad6a47679478b68d79e78e6b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74444
Tested-by: waldhacker <hello@waldhacker.dev>
Tested-by: Björn Jacob <bjoern.jacob@tritum.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: waldhacker <hello@waldhacker.dev>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>