[BUGFIX] Raise phpstan processTimeout
Recently from time to time phpstan ci runs are failing with following
error message:

> Reached internal errors count limit of 50, exiting...
> Internal error: Child process timed out after 600.0 seconds.
> Try making it longer with parallel.processTimeout setting.

As a intermediate solution this timeout should be increased.
Deeper investigation and maybe finding a solution to avoid such
long subprocess run times is the more reasonable way - but to
avoid a lot of test failes, better trade a longer runtime in ci
than a linger time to find a proper fix for it as a first solution.

Resolves: #97490
Releases: main, 11.5
Change-Id: I8f5a23f233024090de32ce4b3272d3800fc0c2ab
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74457
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>