[TASK] Reduce sql queries for page link generation
Generating a page link leads to 2 sql queries.

The first query is about a general lookup and
frontend group access check. The second occurs
through PageRouter::generateUri().

By passing a fully resolved and possibly overlayed
page record (as object) to generateUri() instead of
a pageId the second query can be omitted.

Since generateUri() is public we need to make sure
the page record is completely resolved and overlayed
in order to reuse it. This is done with the new Page object.

Resolves: #97492
Releases: main, 11.5
Change-Id: I307d6e5f53d6581deb494aa123f25bde0a7ff263
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77332
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>