[TASK] Streamline form element labels in form editor
Form elements in the tree component are labelled with the label of the
corresponding form element definition, whereas the stage component uses
the form element's type. This has now been streamlined to always use the
label of the form element definition, since editors might not be
interested in (internal) types anyway.

Resolves: #97507
Releases: main, 11.5
Change-Id: Ife4c52207957db3cfd1cfa1818ccc9628e36b91c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74488
Tested-by: Elias Häußler <e.haeussler@familie-redlich.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: waldhacker <hello@waldhacker.dev>
Reviewed-by: Elias Häußler <e.haeussler@familie-redlich.de>
Reviewed-by: waldhacker <hello@waldhacker.dev>