[TASK] Replace deprecated doctrine/dbal 'getSchemaManager()'
'doctrine/dbal' 3.x deprecated the 'getSchemaManager()'
method in favour of the 'createSchemaManager()' method
on connections. This has been aligned in core with the
doctrine/dbal raise in #96287. However, two occurances
with corresponding todo comments has been overseen.

This patch replace the calls and solves the comments of
the left-overs. Additionally, unrelated wrapper method in
ext:install database checks has aligned to the same name
to avoid confusion when looking for the deprecated method.

Resolves: #97552
Related: #96287
Releases: main
Change-Id: If73e9439acdcbf89db968ba1e6db1c4d60513861
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74524
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>