[TASK] Use PSR-3 compliant logging in ExtendedFileUtility
This change adapts ExtendedFileUtility logs to use
interpolation from PSR-3 instead of %s to identify
variables in log_details properly later-on.

This makes all log messages from ExtendedFileUtility
PSR-3 compliant.

In addition, details_nr is not in use anymore since TYPO3 v6,
so the arbitrary ID is removed when logging.

Also, all messages do not contain a "!" at the end anymore,
as there is no need to emphasize errors through the log message
itself (that's why there is type "Error" for logging).
Same goes for "." in order to be consistent.

Resolves: #97558
Releases: main
Change-Id: I39d474b7c7edd3707d9959b5e750c64f23bcb1ad
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74527
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>