[TASK] CI detection in runTests.sh for phpstan
phpstan parallel execution is limited to a maximum
of 5 processes to be a nice neighbor when executed
during core CI runs. This however leads to longer
execution times when executed on local developer
machines with many available CPU's.

This patch introduces two separated phpstan config
file entrypoints, one for local and one for CI. They
are automatically selected in runTests.sh based on
existence of a CI env variable.

Resolves: #97571
Releases: main, 11.5
Change-Id: I14bf13412de7a2471703e7982dad4030ae6ba735
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74536
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74544