[TASK] Ignore PHPStan errors in GeneralUtilityTest for contract violations
In tests for `GeneralUtility` that explicitly violate the contract
(and test the correct error handling in those cases), PHPStan errors about
the contract violation are not helpful and rather clutter the
PHPStan baseline.

So these errors are now silenced with an annotation in the corresponding
tests, allowing us to focus on PHPStan errors about actual issues.

Note: We only use this annotation in tests, not in production code,
to avoid silencing errors about actual problems.

Resolves: #97574
Releases: main, 11.5
Change-Id: I2586dd77a4e3f9d801e24af68ab29984ad8d02c5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74547
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>