[FEATURE] Provide default queue for notifications
To allow dispatching notifications to the user the easy way, a new global
flash message queue, identified by
`TYPO3\CMS\Core\Messaging\FlashMessageQueue::NOTIFICATION_QUEUE`, is
introduced that takes the flash message and renders it as a notification on
the top-right edge of the backend.

Backend modules based on `TYPO3\CMS\Backend\Template\ModuleTemplate`
automatically gain advantage of this feature.

Resolves: #97595
Releases: main
Change-Id: I0e04f0412117649391c0f50fa4249b7832dded69
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74582
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>