[TASK] CI: Always retry failed jobs once
We restricted the default retry of single jobs a while
ago to do that only for acceptance test related jobs.
This does not work out well: There are simply to many
things that occassionally false-negative due to network,
timing or other issues.
The patch adds general retry:1 for all jobs for mental
sanity.

Resolves: #97597
Related: #96999
Releases: main, 11.5
Change-Id: I408b7ef1b6c6145dc1b5b12bc4550c00dfdaa930
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74585
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>