[TASK] Add support for PHP8.2 to Build/Scripts/runTests.sh
This patch adds the ability to run commands with PHP8.2 using
`Build/Scripts/runTests.sh`. Support is added early to check
which issues may raise up with new major PHP version.

Help text of script is adopted to state the possibilty of the
new PHP version with proper example commands. Additionally a
note and check for currently not suppored xdebug with PHP8.2
is added.

Resolves: #97755
Releases: main, 11.5
Change-Id: I9df13d35278793fba8c5475c8abd602bd1c27896
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74870
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>