[BUGFIX] Toggle inline controls only in own scope
The former approach in #91808 to fix scoped toggling didn't
work with levelLinksPosition "bottom", as the querySelector
could find another control container inside another child
inline field first.

Another solution, which is now possible, works with the
":scope" selector. Only nodes with a direct parent will be
found this way.

In addition, a new helper class is added for the top control
container, as it is wrapped in another div. This enables
toggling of both buttons, in case levelLinksPosition "both"
is set.

Resolves: #97786
Related: #91808
Related: #92382
Releases: main, 11.5
Change-Id: I67146c372d183cd647db847b3e3d83d986585577
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74928
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>