[BUGFIX] Remove unneeded padding from search's form-control
The input field used within the backend search form at the top had an
initial padding of 0px which mispositions the search field until all
JavaScript is loaded because the field will be wrapped in a div
container and thus the responsible CSS rule becomes unused.

To solve this wrong positioning, the affected CSS definition is removed.

Resolves: #97810
Releases: main, 11.5
Change-Id: I01477cd098adcb95fe18467c524a55f8df70da5a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74980
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>