[BUGFIX] Do not use "[Empty]" label for finisher options in FormEngine
With #96478, the default value for overridden finisher options in
FormEngine was set to the translated label "[Empty]" in case the
finisher option was empty. Since applying labels to values does not
really make sense here, the original option value is now set as default
value in FormEngine, ignoring the generated label.

Resolves: #97781
Resolves: #97557
Resolves: #96830
Related: #96478
Releases: 10.4
Change-Id: Ice236994485374a3e5399926df901da15cb8d991
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74999
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>