[BUGFIX] Cast folder name in tree provider to string
The method FileStorageTreeProvider->getSubfoldersRecursively() iterates
an array of subfolders and prepares folder information based on the
iterated data.

This functionality has a minor flaw, thanks to PHP, in case a folder
is named with numbers only, PHP treats array indices as integer if they
can be parsed as such, leading to TypeErrors as the method
prepareFolderInformation() expects a nullable string.

The passed subfolder name is now cast to string to solve this issue.

Resolves: #97851
Releases: main, 11.5
Change-Id: I3357121c4892d6063b6714e4e7b436b3ebdd1802
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75028
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>