[BUGFIX] Adapt `Status->getDetailedSystemStatus()` to use StatusRegistry
In the introduction of the StatusRegistry it was forgotten to adapt one
method in the Status class. The adaption renders the property
`statusProviders` now obsolete, and it is therefore removed.

In addition, a variable name in the `getSystemStatus` method is
adjusted and the null coalescing assignment is used.

Resolves: #97883
Related: #97320
Releases: main
Change-Id: Ie71dc7fad56f2a0545f1446de1f420c6f031cf2d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75087
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>