[!!!][FEATURE] Use LanguageAspect in QuerySettings for Extbase
This change adapts the QuerySettings functionality
for Extbase in v12 to allow to work with a LanguageAspect
which contains all information regarding the language ID
and the overlay mode.

For this reason, the following method definitions are removed
within Extbase's QuerySettingsInterface

- setLanguageOverlayMode()
- getLanguageOverlayMode()
- setLanguageUid()
- getLanguageUid()

However, it is still possible to use the existing methods
on the QuerySettings implementation of TYPO3
for TYPO3 v12.

It is now recommended to use ->getLanguageAspect()
and ->setLanguageAspect().

This change is a pre-requisite to extend language
handling across TYPO3 and Extbase.

Resolves: #97926
Releases: main
Change-Id: I38a685fcfdb98d5ec3c82b02e6cc78df12062fa0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75093
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>