[BUGFIX] Show rollback link in history again
The rollback link for a single field change is
now displayed again in the rollback view of the
history module by using the correct variable in
the template.

Additionally, an unintended change in #91148 is
resolved, which would also display the rollback
link in the normal history list view.

Resolves: #97901
Releases: main, 11.5
Change-Id: I2606d0585a5f1d196adab5f83340ff3033dced1c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75071
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>