[BUGFIX] Allow changing FF sections in new version overlays
Creating a version overlay of a record that has flexform
sections and moving or deleting some of these sections
at the same time fails in DataHandler due to a wrong uid
lookup for the section command array.

The patch fixes this scenario in the DataHandler flexform
processing code and leaves an according explanation comment.

Resolves: #73761
Releases: main, 11.5
Change-Id: Ib1c910c59a2ebfe9b7545cabb258f44bac2cd273
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75116
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Tina Westner <tina.westner@bgm-gmbh.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Tina Westner <tina.westner@bgm-gmbh.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>