[BUGFIX] Disable asset concatenation/compression in backend
Scenario: Given is a cluster of containers/servers that run
multiple instances of the same TYPO3 instance. A load balancer
round robins requests to those instances.

Problem: A backend user logs in and gets a broken backend.
No CSS and JavaScript is loaded. Requests to those
resources return HTTP 404 errors.

Analysis: TYPO3 concatenates assets on the fly per backend
module when they are requested and saves them to a local
temporary folder. As requests are routed to different
containers on every request, the probability of hitting
an container/server where the resulting asset is not yet
generated, is very high (basically always). The only way
to circumvent this, is to put the backend in debug mode,
which has other unwanted side effects.

Solution: Deactivate asset compression/concatenation
altogether for the backend. As most servers deliver
HTTP2/3 nowadays, the performance impact should be minimal.
By skipping unnecessary processing and allowing browser
caching, performance could be even increased.

Resolves: #97936
Releases: main, 11.5
Change-Id: Ic696f4a3618e5bed3388ab0262ce2867d334723f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75175
Tested-by: core-ci <typo3@b13.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>