[BUGFIX] Skip resolving backpath for already absolute paths
In case GU::createVersionNumberedFilename() receives
a file with an absolute path, no back path should be resolved
since the resolved path will always be invalid, which led
to no version number got added to the filename anymore.

This is now fixed by checking for the path being absolute.

Resolves: #97939
Releases: main, 11.5, 10.4
Change-Id: I5bb0150fa27b8c9c1af2aa99bd8baacd55889245
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75141
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>