[BUGFIX] Add sys_language_uid for PageSlugCandidateProvider
When using routing and custom enhancers to find out
what page translation was exactly used to build
special logic for Page resolving, the two
language-related properties "sys_language_uid"
and "l18n_cfg" are now loaded (and thus, added to the
"_page" property of a Route) as well.

This is useful as l10n_parent is already loaded anyways.

Resolves: #98042
Releases: main, 11.5, 10.4
Change-Id: I23f964d5d22f27a9affebfeae87906d9549efda0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75311
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>