[BUGFIX] Fix relationship import with sql strict_mode
When sql strict mode is enabled on the database server,
if a value has a wrong data type for the database column,
the INSERT or UPDATE query will fail.
This happened during a t3d import: relationships are lost.

This patch fixes the issue by using the TCA default value for the field.

Resolves: #89103
Releases: main, 11.5
Change-Id: I7e995f10359c6935f6b9a50b1150c31e63937ab9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75391
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>