[!!!][TASK] Remove support for concat/compress of JS+CSS in BE
This change removes the support for supporting concatenation and
compression of CSS + JS assets in TYPO3 Backend.

This feature was actually introduced back in TYPO3 v4.x because of
limitations of Internet Explorer (max of 31 CSS files per HTML page),
and before the introduction of bundlers + compressors for TYPO3 Core,
as well as HTTP/2 support and optimized caching.

For this reason, all the functionality within PageRenderer and
ResourceCompressor is removed in TYPO3, without any option to
activate this.

TYPO3 already bundles CSS in one single file, and JS is nowadays
loaded with bundlers, and loaders within JS modules in TYPO3 Backend.

For TYPO3 Frontend, this functionality is still available for
the time being.

Resolves: #98100
Releases: main
Change-Id: Id45714d4660080da6f4a1d9d6ae277930c0bbc48
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75389
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>