[TASK] Enable PHP8.2 functional testing in CI workflows
This patch shifts some functional test executions in
CI workflows to PHP8.2. Thus enabling test coverage
for new PHP8.2 version in a early stage.

For nightly we only can active it for `locked` runs,
as prophecy/prophecy is not allowed to be installed
with PHP8.2 currently.
See: https://github.com/phpspec/prophecy/issues/556

Backport to 11.5 is skipped for now due some issues
in "doctrine/dbal 2.x" with PHP8.2.

Resolves: #98109
Releases: main
Change-Id: I02c4dc4bd3cdd589330b038463ed58a384e3d70d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75404
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>