[BUGFIX] Reflect size property to attribute in icon-element
When using the custom element <typo3-backend-icon>
with a JS framework, that renders the DOM (e.g. Vue),
it can happen, that they set the properties of the custom element
instead of an attribute.

Since Lit does not reflect these properties to attributes by default,
but the icon-element relies on CSS set on the host element that depends
on the size attribute, the icon is not rendered in the correct size.

See: https://web.dev/custom-elements-best-practices/
on why reflecting properties to attributes might be a good idea
in some cases.

Releases: main, 11.5
Resolves: #98121
Change-Id: I8435946ad6e6d92b8ec50dd7e289279261d43d5a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75414
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>