[BUGFIX] Restore module menu collapse behaviour
With the removal of the Font Awesome icon provider in #98089 the
corresponding scss/css styling definition for the module name
has been removed. This made it impossible to fully collapse the
module menu to icons and to a smaller size.

Bootstrap 5 replaced `.sr-only` with `.visually-hidden`. However,
this does not work for the module menu as the menu entry names
are hidden in expanded mode, thus making this a non-usable
option.
See: https://getbootstrap.com/docs/5.0/migration/#helpers
and: https://getbootstrap.com/docs/5.0/helpers/visually-hidden/

This patch re-adds scss/css styling directly for the module
menu without using the `@sr-only` or the new `.visually-hidden`
definition from bootstrap to restore the module menu collapse
behaviour.

Used commands:

> Build/Scripts/runTests.sh -s clean
> Build/Scripts/runTests.sh -s buildCss

Resolves: #98138
Related: #98089
Releases: main
Change-Id: I200e9a5324be93fdc5119c956cfe432a1ecbc81d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75436
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>