[BUGFIX] Always load label for edit notification
If a record is edited via single field mode (aka "columnsOnly"), the
label field as defined via TCA may be missing, rendering "[No title]" in
the notification message.

Before compiling the notification message, it is now checked whether the
label exists already. If not, the label of the record is fetched from
the database, if such label field is configured via TCA at all.

Resolves: #98142
Releases: main
Change-Id: I2babfcae354b1fad15c653e61449353cd2efb8a4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75443
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>