[TASK] Remove jQuery from backend login
This patch aims to remove jQuery from the TYPO3 backend login.
Since some jQuery magic was used to dynamically adjust the height of
the carousel with a nice animation, which is not so easy to solve in
pure CSS and JavaScript, the carousel now has a fixed height.

To make the approach a little bit more appealing and usable for mobile
users who usually don't see a scrollbar, the top and bottom area of the
carousel container are faded if text is overflown to indicate a
scrollable area.

To make sliding to the next slide in the carousel less "jumpy", smooth
scrolling is explicitly enabled here.

Resolves: #98157
Releases: main
Change-Id: I08ecd847d29696fd25f6330cee089a55a03102c0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75473
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>