[BUGFIX] Avoid open_basedir errors when referencing absolute URLs
Backend modules which use absolute paths (which do not
exist) should not raise an exception / error when
using versionNumberInFilename checks.

This happens e.g. with EXT:solr backend modules
since #97939 was fixed.

In the long-term extensions should define proper
paths, so they also work in TYPO3 installations within
a subfolder but for the time being,
this change helps to solve the problem.

Resolves: #98152
Related: #98106
Related: #97939
Releases: main, 11.5, 10.4
Change-Id: I071f2e63d4423d157f0f060a1295cb145474ee9d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75494
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>