[!!!][TASK] Drop backend login interface selector
Previous TYPO3 installations allowed to configure an interface selector
in the backend login that gave the authenticating backend user the
possibility to choose whether to get redirected to frontend or backend
by configuring
`$GLOBALS['TYPO3_CONF_VARS']['BE']['interfaces'] = 'backend,frontend'`.

If only one option was configured, the redirect to either backend or
frontend was enforced, where `backend` was the default configuration.

This feature was meaningful once TYPO3 shipped EXT:feedit, but was
conceptually broken ever since, as the matter of fact a TYPO3
installation can contain multiple site roots was overseen and a user
may get redirected to the wrong frontend. Also, if EXT:adminpanel is not
installed, there is no one-click solution to access the TYPO3 backend.

Resolves: #98179
Releases: main
Change-Id: Ic759e6c1016d70c07dc8862a0b8c21b14c3d904b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75487
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>