[TASK] Allow `Element` and `DocumentFragment` in now stricter Modal API
The configuration definition of `content` in the Modal API is extended
to accept objects of types `Element` and `DocumentFragment` as well,
which allows passing arbitrary DOM nodes to the API, if TypeScript is
used. Since jQuery is still the main actor in the Modal API, no further
adjustments are required for now as jQuery can consume the added types.

In the same run, the API definition got stricter as `Modal.advanced()`
allowed to pass any arbitrary configuration not covered by the internal
`Configuration` interface.

Resolves: #98164
Releases: main, 11.5
Change-Id: Ibf7cc836fcffa870a1eb4c5d4615c44e046cf192
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75488
Tested-by: core-ci <typo3@b13.com>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>