[BUGFIX] Avoid invalid property access when persisting extbase objects
With #95819 the detection of Domain Object properties was changed to use
ClassSchema in order to support uninitialized properties is models.
This was not possible with `DomainObjectInterface::_getProperties()`
which used `get_object_vars()` under the hood.

However, with mentioned change, other properties which had been
ignored by `get_object_vars()` were now propagated to persistence,
even if those properties were not configured in TCA. The change in
behavior is considered a regression and is now addressed:

The retrieval of the property value is delayed until after it is
ensured that the property is a persistable property (which asserts
that we are not reading from private properties).

Resolves: #98148
Related: #95819
Releases: main, 11.5
Change-Id: I94d41715a49953045d32361a00bc274c15112a1e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75541
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>